[Chore/#63] viewModel들 폴더에 넣기
