+++ app/src/main/java/com/sopt/umbba_android/presentation/home/viewmodel/InviteCodeDialogFragmentViewModel.kt
@@ -1,4 +1,4 @@
-package com.sopt.umbba_android.presentation.home
+package com.sopt.umbba_android.presentation.home.viewmodel
 
 import androidx.lifecycle.ViewModel
