+++ app/src/main/java/com/sopt/umbba_android/presentation/onboarding/SelectFamilyActivity.kt
@@ -6,6 +6,7 @@ import android.view.View
 import androidx.activity.viewModels
 import com.sopt.umbba_android.R
 import com.sopt.umbba_android.databinding.ActivitySelectFamilyBinding
+import com.sopt.umbba_android.presentation.onboarding.viewmodel.SelectFamilyViewModel
 import com.sopt.umbba_android.util.binding.BindingActivity
 
 class SelectFamilyActivity :
