+++ app/src/main/java/com/sopt/umbba_android/presentation/qna/AnswerActivity.kt
@@ -4,12 +4,11 @@ import android.os.Bundle
 import android.util.Log
 import android.view.View
 import androidx.activity.viewModels
-import androidx.lifecycle.viewModelScope
 import com.sopt.umbba_android.R
 import com.sopt.umbba_android.databinding.ActivityAnswerBinding
+import com.sopt.umbba_android.presentation.qna.viewmodel.AnswerViewModel
 import com.sopt.umbba_android.util.ViewModelFactory
 import com.sopt.umbba_android.util.binding.BindingActivity
-import timber.log.Timber
 
 class AnswerActivity : BindingActivity<ActivityAnswerBinding>(R.layout.activity_answer),
     View.OnClickListener {
