+++ app/src/main/java/com/sopt/umbba_android/presentation/qna/ConfirmAnswerDialogFragment.kt
@@ -4,20 +4,17 @@ import android.graphics.Color
 import android.graphics.drawable.ColorDrawable
 import android.os.Bundle
 import android.util.Log
-import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.view.WindowManager
 import android.widget.Toast
-import androidx.activity.viewModels
 import androidx.fragment.app.DialogFragment
 import androidx.fragment.app.viewModels
-import com.sopt.umbba_android.R
 import com.sopt.umbba_android.data.model.request.AnswerRequestDto
 import com.sopt.umbba_android.databinding.FragmentConfirmAnswerDialogBinding
+import com.sopt.umbba_android.presentation.qna.viewmodel.ConfirmAnswerDialogFragmentViewModel
 import com.sopt.umbba_android.util.ViewModelFactory
-import timber.log.Timber
 
 class ConfirmAnswerDialogFragment : DialogFragment() {
