+++ app/src/main/java/com/sopt/umbba_android/presentation/qna/QuestionAnswerActivity.kt
@@ -9,7 +9,7 @@ import androidx.core.content.ContextCompat
 import com.sopt.umbba_android.R
 import com.sopt.umbba_android.data.model.response.QuestionAnswerResponseDto
 import com.sopt.umbba_android.databinding.ActivityQuestionAnswerBinding
-import com.sopt.umbba_android.presentation.home.InviteCodeDialogFragment
+import com.sopt.umbba_android.presentation.qna.viewmodel.QuestionAnswerViewModel
 import com.sopt.umbba_android.util.ViewModelFactory
 import com.sopt.umbba_android.util.binding.BindingActivity
