+++ app/src/main/res/layout/activity_answer.xml
@@ -7,7 +7,7 @@
 
         <variable
             name="vm"
-            type="com.sopt.umbba_android.presentation.qna.AnswerViewModel" />
+            type="com.sopt.umbba_android.presentation.qna.viewmodel.AnswerViewModel" />
 
         <variable
             name="titleText"
